Moved Laser class to own file
Added author
