+++ src/test1/visuals.py
@@ -1,18 +1,4 @@
-import pygame
-
-
-class Laser:
-    def __init__(self, colors=('red', 'red2', 'white')):
-        self.glow = [pygame.color.THECOLORS[color] for color in colors]
-
-    def draw_laser(self, surface, start, stop, magnitude):
-        width = magnitude
-        for color in self.glow:
-            if width <= 0:
-                break
-            pygame.draw.line(surface, color, start, stop, width)
-            width -= 2
-
+import Laser as lsr
 
 class Visuals:
 
@@ -22,8 +8,8 @@ class Visuals:
 
     whiteColor = (255, 255, 255)
     bgColor = (0, 0, 0)
-    redLaser = Laser(['red4', 'red', 'white'])
-    blueLaser = Laser(['blue', 'blue2', 'white'])
+    redLaser = lsr.Laser(['red4', 'red', 'white'])
+    blueLaser = lsr.Laser(['blue', 'blue2', 'white'])
 
     def __init__(self):
         pass
